requirements.txt
fastapi
fastapi[all]
uvicorn
python-dotenv
